requirements.txt
ultralytics>=8.3.0
opencv-python>=4.10.0
pyyaml>=6.0.2
tqdm>=4.66.0
pytest>=8.0.0
